Deprecated node indexer in favour of nats

+++ etc/schema/types/block_result.sql
@@ -15,7 +15,7 @@ CREATE TYPE block_result AS (
     "totalDifficulty" int2,
     "extraData" bytea,       -- 0 bytes
     "size" int4,
-    "gasLimit" int8,
+    "gasLimit" numeric,
     "gasUsed" int8,
     "timestamp" int4,
     "mixHash" hash
